requirements.txt
pykrx>=1.2.4
pandas>=2.0
numpy>=1.24
python-telegram-bot>=20.0
APScheduler>=3.10
matplotlib>=3.7
ta>=0.11
setuptools<71
python-dotenv>=1.0
pyyaml>=6.0
